requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
marshmallow==3.20.1
flask-cors==4.0.0
